move utils to front folder as it is in git ignore
